+++ mixins/firmware/getFirmwareVersionMixin.js
